Refactor duplicated map graph distance logic into a common module
Extracted duplicated `getDistanceToMap` and `resolveOutdoorMapId` logic from gen1, gen2, and gen3 graphs into a unified `common.ts` module to improve maintainability and DRYness.
